Fixed Log Manager for the game Hex
The Log Manager does not Work with Applets, because Applets set the
default working directory to: "C:\WINDOWS\System32".

If you want to use the Log Manager just start Hex with
"LaunchTrainHex.java". The Hex Tile also needs to be serializable.

If you need the Applet for whatever reason, let me know and maybe I can
find a fix.